requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
asgiref==3.11.0
attrs==25.4.0
blinker==1.9.0
cachetools==7.0.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
cloudinary==1.44.1
dj-database-url==3.1.0
Django==6.0.2
django-cloudinary-storage==0.3.0
django-cuser==2017.3.16
django-modeltranslation==0.19.19
django-tinymce==5.0.0
Flask==3.1.2
frozenlist==1.8.0
gunicorn==25.0.0
idna==3.11
ipinfo==5.4.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
multidict==6.7.1
packaging==26.0
pillow==12.1.0
propcache==0.4.1
pytz==2025.2
requests==2.32.5
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
Werkzeug==3.1.5
wheel==0.45.1
whitenoise==6.11.0
yarl==1.22.0
